pyproject.toml
[tool.poetry]
name = "Vinter tech challenge"
version = "0.1.0"
description = "Vinter tech challenge for Senior Quantitative Analyst position."
authors = ["Lukas Bures <burylukas@seznam.cz>"]

[tool.poetry.dependencies]
python = "^3.9"
backtrader = { git = "https://github.com/mementum/backtrader.git", rev = "master" }
pandas = "^1.4.3"
starlette = "^0.20.4"
matplotlib = "^3.5.3"
backtrader-plotting = "^2.0.0"
QuantStats = "^0.0.59"
pyfolio = "^0.9.2"
python-binance = "^1.0.16"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
